+++ src/Rib.hs
@@ -2,11 +2,8 @@
 module Rib
   ( module Rib.App,
     module Rib.Shake,
-    Source,
+    module Rib.Target,
     SourceReader,
-    sourcePath,
-    sourceVal,
-    sourceUrl,
     MMark,
     Pandoc,
   )
@@ -14,6 +11,7 @@ where
 
 import Rib.App
 import Rib.Source
+import Rib.Target
 import Rib.Parser.MMark (MMark)
 import Rib.Parser.Pandoc (Pandoc)
 import Rib.Shake