+++ CHANGELOG.md
@@ -11,8 +11,9 @@
   - `ghcSyntaxHighlighter`: we already have `skylighting` (which supports more parsers than Haskell)
   - `obfuscateEmail`: requires JS, which is not documented.
 - API changes:
+  - Removed `buildHtmlMulti`, `buildHtml` and `Source` type.
+  - Introduced `Target` type, and `loadTarget`/`writeTarget` functions.
   - Introduced `forEvery` to run a Shake action over a pattern of files when they change.
-  - Removed `buildHtmlMulti` (use `forEvery` with `buildHtml` instead)
   - Exposed `Rib.Shake.writeFileCached`
 - Bug fixes
   - #95: Fix Shake error `resource busy (file is locked)`